+++ indico/modules/events/editing/operations_test.py
@@ -11,22 +11,27 @@ from indico.modules.events.editing.models.editable import Editable, EditableType
 from indico.modules.events.editing.models.revisions import EditingRevision, RevisionType
 
 
-def test_cannot_undo_review_old_rev(dummy_contribution, dummy_user):
+@pytest.mark.usefixtures('request_context')
+def test_cannot_undo_review_old_rev(db, dummy_contribution, dummy_user):
     from indico.modules.events.editing.operations import InvalidEditableState, undo_review
     editable = Editable(contribution=dummy_contribution, type=EditableType.paper)
     rev1 = EditingRevision(editable=editable, user=dummy_user, type=RevisionType.ready_for_review)
     rev2 = EditingRevision(editable=editable, user=dummy_user, type=RevisionType.needs_submitter_confirmation)
     rev3 = EditingRevision(editable=editable, user=dummy_user, type=RevisionType.changes_rejection)
+    db.session.flush()
+
     with pytest.raises(InvalidEditableState):
         undo_review(rev1)
     with pytest.raises(InvalidEditableState):
         undo_review(rev2)
     undo_review(rev3)
+
     assert not rev1.is_undone
     assert not rev2.is_undone
     assert rev3.is_undone
 
 
+@pytest.mark.usefixtures('request_context')
 @pytest.mark.parametrize(('type', 'ok'), (
     (RevisionType.new, False),
     (RevisionType.ready_for_review, False),
@@ -41,6 +46,7 @@ def test_cannot_undo_review_old_rev(dummy_contribution, dummy_user):
 def test_can_undo_review(db, dummy_contribution, dummy_user, type, ok):
     from indico.modules.events.editing.operations import InvalidEditableState, undo_review
     editable = Editable(contribution=dummy_contribution, type=EditableType.paper)
+    EditingRevision(editable=editable, user=dummy_user, type=RevisionType.ready_for_review)
     rev = EditingRevision(editable=editable, user=dummy_user, type=type)
     db.session.flush()
 
@@ -53,10 +59,10 @@ def test_can_undo_review(db, dummy_contribution, dummy_user, type, ok):
 
     db.session.expire(editable)  # so a deleted revision shows up in the relationship
     if type == RevisionType.reset:
-        assert len(editable.revisions) == 0
-    else:
         assert len(editable.revisions) == 1
-    if ok or type == RevisionType.reset:
-        assert len(editable.valid_revisions) == 0
     else:
+        assert len(editable.revisions) == 2
+    if ok or type == RevisionType.reset:
         assert len(editable.valid_revisions) == 1
+    else:
+        assert len(editable.valid_revisions) == 2