+++ CHANGES.rst
@@ -48,6 +48,7 @@ Improvements
   :user:`kewisch`)
 - Allow editors to edit their review comments on editables (:pr:`6008`)
 - Auto-linking of patterns in minutes (e.g. issue trackers, Github repos...) (:pr:`5998`)
+- Log editor actions in the Editing module (:pr:`6015`)
 
 Bugfixes
 ^^^^^^^^